// gitlabdockercompose-main/app/types/errors/messages.ts
export const enum ErrorMessages {
  E401 = 'You are not authorized to access this resource',
  E1001 = 'Error fetching data from resource',
  E1002 = 'Error fetching data from resource with provided identifier',
  E1003 = 'Error creating resource',
  E1004 = 'Error updating resource with provided identifier',
  E1005 = 'Error deleting resource with provided identifier',
  E1006 = 'Unable to find the requested resource with provided identifier',
  E1007 = 'Validation failed for the provided input',
  E1008 = 'Error uploading files to resource with provided identifier',
  E1009 = 'No changes to be saved',
  E1010 = 'Email address already exists',
  E1011 = 'Phone number already exists',
  E1012 = 'Account has been deactivated',
  E1013 = 'Error authenticating your account',
  E1014 = 'Error performing the requested action',
  E2001 = 'Unable to send notification via the given channel',
  E6301 = 'Namespace is required while creating redis instances',
}